added flexibility
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@5093 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ mlab/amoebaGen/plotAmoebaR2.m
@@ -69,7 +69,7 @@ image_filename = ...
     [image_file_dir, fname, suffix_zeros, num2str(trial)];
 % image_filename = ...
 %     [image_file_path, fname, suffix_zeros, num2str(trial)];
-savefile2(image_filename, image);
+savefile2(image_filename, uint8(image));
 
 global plot_amoeba2D fh_amoeba2D
 if plot_amoeba2D
